first commit with code for assaults
I binned the assaults into regions and exported the information into csv. I am a little confused by how it output because the program had to condense hours and days into one dimension so it could be viewed in a 2D spreadsheet. I think it stored as each hour for sunday then each hour for monday etc... which is basically hour of the week, but it could have done each day at midnight, then each  day at 1 am, then 2am... I will play around with it soon and see if I figure it out